backend/requirements.txt
fastapi[standard]
PyJWT
passlib[bcrypt]
sqlmodel
alembic